Flatten key type hierarchy
We weren't actually using polymorphism and this makes cloning easier.
